+++ ScissorGlue/!_Source/BackGround/BackGround.h
@@ -1,15 +1,14 @@
+
 #pragma once
-/*****************************************************************//**
- * \file   BackGround.h
- * \brief  ”wŒiˆ—
- * 
- * \author ‚ß‚´‚µ
- * \date   July 2022
- *********************************************************************/
 
+// ”wŒiƒNƒ‰ƒX
+class BackGround {
+public:
+	BackGround();
+	~BackGround();
 
-class BackGround
-{
+	void	Draw();
 
-	aaa
+private:
+	int		_cg;		// ‰æ‘œ
 };
