--- ScissorGlue/!_Source/Map/BackGround.h
@@ -1,23 +1,0 @@
-#pragma once
-/*****************************************************************//**
- * \file   BackGround.h
- * \brief  ”wŒiˆ—
- * 
- * \author ‚ß‚´‚µ
- * \date   July 2022
- *********************************************************************/
-
-
-
-
- // ”wŒiƒNƒ‰ƒX
-class BackGround {
-public:
-	BackGround();
-	~BackGround();
-
-	void	Draw();
-
-private:
-	int		_cg;		// ‰æ‘œ
-};
