+++ ScissorGlue/!_Source/Map/MapChips.cpp
