+++ ScissorGlue/!_Source/Map/MapChips.h
